refactor(Panel): BaseディレクトリをPanelディレクトリに統合
- Base ディレクトリ → Panel ディレクトリに変更
- Panel, BaseColumn を Panel ディレクトリに集約
- 全てのインポートパスを Panel ディレクトリからに変更
- panelClassNameGenerator を Panel ディレクトリからエクスポート
