+++ packages/smarthr-ui/src/components/InformationPanel/InformationPanel.tsx
@@ -14,7 +14,6 @@ import { type VariantProps, tv } from 'tailwind-variants'
 
 import { useObjectAttributes } from '../../hooks/useObjectAttributes'
 import { useIntl } from '../../intl'
-import { Panel, type PanelElementProps } from '../Base'
 import { Button } from '../Button'
 import { Heading, type HeadingTagTypes } from '../Heading'
 import {
@@ -27,6 +26,7 @@ import {
   WarningIcon,
 } from '../Icon'
 import { Sidebar } from '../Layout'
+import { Panel, type PanelElementProps } from '../Panel'
 
 type ObjectHeadingType = {
   text: ReactNode
