+++ packages/smarthr-ui/src/components/Panel/index.ts
@@ -1,3 +1,3 @@
-export { Panel } from './Panel'
+export { Panel, panelClassNameGenerator } from './Panel'
 export type { ElementProps as PanelElementProps } from './Panel'
 export { BaseColumn } from './BaseColumn'
