+++ packages/smarthr-ui/src/components/Panel/stories/BaseColumn.stories.tsx
@@ -1,13 +1,13 @@
 import { backgroundColor } from '../../../tailwind'
 import { defaultBackgroundColor as backgroundColorValue } from '../../../themes'
 import { Stack } from '../../Layout'
-import { baseClassNameGenerator } from '../Base'
 import { BaseColumn } from '../BaseColumn'
+import { panelClassNameGenerator } from '../Panel'
 
 import type { Gap } from '../../../types'
 import type { Meta, StoryObj } from '@storybook/react-webpack5'
 
-const basePadding = Object.keys(baseClassNameGenerator.variants.paddingBlock)
+const basePadding = Object.keys(panelClassNameGenerator.variants.paddingBlock)
   // Tシャツサイズは後方互換性のために残しており、できるだけ使われたくない
   .filter((v) => !isNaN(Number(v)))
   .sort() as Gap[]
