+++ packages/smarthr-ui/src/components/Panel/stories/Panel.stories.tsx
