+++ packages/smarthr-ui/src/components/Panel/stories/VRTPanel.stories.tsx
